# control-plane/internal/apispec/openapi.yaml
openapi: 3.1.0
info:
  title: Hive API
  version: 0.5.0
  description: 'Hive is a Swarm-native platform control plane. This spec is the single

    source of truth for all API contracts. Every request body, response, and

    error shape is defined here.

    '
servers:
- url: /api/v1
  description: Control plane API v1
components:
  schemas:
    ErrorResponse:
      description: ErrorResponse schema.
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error message.
      example:
        code: bad_request
        message: invalid payload
    StatusResponse:
      description: StatusResponse schema.
      type: object
      additionalProperties: false
      required:
      - status
      properties:
        status:
          type: string
          description: Current status.
      example:
        status: ok
...
    IdResponse:
      description: IdResponse schema.
      type: object
      additionalProperties: false
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier.
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
...
    Stack:
      description: Stack schema.
      type: object
      additionalProperties: false
      required:
      - id
      - projectId
      - name
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier.
        projectId:
          type: string
          format: uuid
          description: Project ID.
        name:
          type: string
          description: Display name.
        composeContent:
          type: string
          description: Compose content.
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601).
    CreateStackRequest:
      description: CreateStackRequest schema.
      type: object
      additionalProperties: false
      required:
      - projectId
      - name
      - composeContent
      properties:
        projectId:
          type: string
          format: uuid
          description: Project ID.
        name:
          type: string
          description: Display name.
        composeContent:
          type: string
          description: Compose content.
      example:
        projectId: 550e8400-e29b-41d4-a716-446655440000
        name: monitoring
        composeContent: "version: \"3.8\"\nservices:\n  prometheus:\n    image: prom/prometheus\n"
    UpdateStackRequest:
      description: UpdateStackRequest schema.
      type: object
      additionalProperties: false
      required:
      - composeContent
      properties:
        composeContent:
          type: string
          description: Compose content.
      example:
        composeContent: "version: \"3.8\"\nservices:\n  grafana:\n    image: grafana/grafana\n"
...
    Service:
      description: Service schema.
      type: object
      additionalProperties: false
      required:
      - id
      - name
      properties:
        id:
          type: string
          description: Unique identifier.
        name:
          type: string
          description: Display name.
        image:
          type: string
          description: Docker image reference.
        replicas:
          type: integer
          description: Number of replicas.
        running:
          type: integer
          description: Running.
        labels:
          type: object
          additionalProperties:
            type: string
          description: Key-value labels.
...
    Settings:
      type: object
      additionalProperties: true
      description: Key-value map of global application settings.
...
    StackActionResponse:
      description: Result of a stack start/stop/restart action.
      type: object
      additionalProperties: false
      required:
      - status
      - replicas
      properties:
        status:
          type: string
          description: Action result status.
        replicas:
          type: integer
          description: Replica count every stack service was scaled to.
      example:
        status: ok
        replicas: 1
...
  parameters:
    IdParam:
      description: IdParam schema.
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadGateway:
      description: Upstream (agent or Docker) unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      description: bearerAuth schema.
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /services:
    get:
      operationId: listServices
      summary: List Swarm services
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Service list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
      description: Retrieve services
...
  /stacks:
    get:
      operationId: listStacks
      summary: List stacks
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Stack list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Stack'
      description: Retrieve stacks
    post:
      operationId: createStack
      summary: Create stack
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStackRequest'
      responses:
        '201':
          description: Stack created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      description: Create stacks
  /stacks/{id}:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getStack
      summary: Get stack
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Stack
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Retrieve stacks id
    put:
      operationId: updateStack
      summary: Update stack
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStackRequest'
      responses:
        '200':
          description: Stack updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Update stacks id
    delete:
      operationId: deleteStack
      summary: Delete stack
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Stack removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Delete stacks id
  /stacks/{id}/deploy:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    post:
      operationId: deployStack
      summary: Deploy stack
      security:
      - bearerAuth: []
      responses:
        '202':
          description: Stack deploy enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Create stacks id deploy
      requestBody:
        description: Empty request body.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
  /stacks/{id}/start:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    post:
      operationId: startStack
      summary: Start stack
      description: Scales every service in the stack back to one replica.
      security:
      - bearerAuth: []
      requestBody:
        description: Empty request body.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Stack started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /stacks/{id}/stop:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    post:
      operationId: stopStack
      summary: Stop stack
      description: Scales every service in the stack to zero replicas.
      security:
      - bearerAuth: []
      requestBody:
        description: Empty request body.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Stack stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /stacks/{id}/restart:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    post:
      operationId: restartStack
      summary: Restart stack
      description: Restarts every service in the stack by scaling it back to one replica.
      security:
      - bearerAuth: []
      requestBody:
        description: Empty request body.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Stack restarted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /settings:
    get:
      operationId: getSettings
      summary: Get global settings
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Settings map
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    $ref: '#/components/schemas/Settings'
      description: Retrieve settings
    put:
      operationId: putSettings
      summary: Update global settings
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
      description: Update settings
...